Dynamically populate recommendations + started maps fragment
Summary: The recommendations are dynamically populated based on selections. The solution is not optimal: in order to be able to search Yelp by category, I had to use the aliases instead of the full category names. We should switch to a more robust model for categories (like an actual Category model instead of a String that holds the alias). Now the app shows the aliases of the categories instead of the full names. This also causes the icons not to display. Not too difficult to fix hopefully.

Test Plan: {F133417926}

Reviewers: madjangba, kiira, kychow

Reviewed By: kychow

Differential Revision: https://phabricator.intern.facebook.com/D8945048

Tasks: T31833570

Signature: 8945048:1532376393:35c98a851986cdcd95f9ab0eeaba386d4b17047e