More recommendations
Summary: More recs appear now. If we want to make even more appear, that's just one line we need to change (I'm still hard-coding the amount of recs). The good news is that there is now logic to handle the case when we reach the end of the business list and haven't yet finished populating the recs (this only happens when multiple places belong to two categories).

Test Plan: N/A

Reviewers: madjangba, kiira, kychow

Reviewed By: madjangba

Differential Revision: https://phabricator.intern.facebook.com/D9166141

Tasks: T32250431

Signature: 9166141:1533578688:37bc15359986305788c4dc9af506e32b93b62f64