can click on item in itinerary to view details
Summary:
Created OnItemViewClickedListener interface in Itinerary fragment to send data to details fragment. The interface is implemented in ItineraryMaps and MainActivity; implemented method (inflateDetails) for the interface is the exact same method as the OnSelectedListener in the Recs fragment to avoid code duplication.

Meant to create a separate diff, but this now includes the frontend for share button. Changed the layout for the itinerary fragment from a linear layout to a relative layout to allow for the share button to be added to the top right of the fragment. Share button implements an OnClickListener to change the color from black to blue, and eventually brings up the share dialogue(?) box.

Test Plan:
{F135280958}
{F135354367}

Reviewers: j053, madjangba, kychow

Reviewed By: madjangba

Differential Revision: https://phabricator.intern.facebook.com/D9193597

Tasks: T32486594, T32471725

Signature: 9193597:1533669665:d279af43589ef9e6da5746c1a7b259dbb144893a